# triangulate-spade: small optimization, potentially up to a factor of 1/2